Milestone 2 70%
Big day yesterday and today without a commit… worked on many features
these past 2 days.. Couple Folder changes occurred and messed up a few
files… but got the following working after a little file fixing. I have
the background Colors changing at this point. I have access to the
camera and the gallery for backgrounds. I created Pickers for the
times, and got the toggle on/off switch to work with the icons on the
bottom. I also had to implement a "save Settings" to make sure those
are stored for when the user comes back. to the app. Changed the fonts
to the ones in the mock up as well. The transistion modes are not quite
there yet.